Doc and doctests (#11)
* change dir tree

* Add init in submodules

* sphinx doc files

* Switched to micromind name

* change phinet into micromind

* Moved PhiNet convolutional block in networks

* Changed file structure

* updated init

* fixed import

* fixed bug on padding

* removed debug breakpoint

* fixed linters

* Changed rst for micromind.networks

* fixed doc after restructuring

* black fix

* Update with doctest docstrings on each example

run it with "make doctest" into docs diretory.
model.from_checkpoint("checkpoint.pt") test creates file "checkpoint.pt" into docs directory, consider removing such function.

* Update with fixed doctest on each example

* Update README.md

---------
