+++ docs/source/index.rst
@@ -1,0 +1,21 @@
+.. micromind documentation master file, created by
+   sphinx-quickstart on Fri Feb 24 15:07:22 2023.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+Welcome to micromind's documentation!
+=====================================
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
+   modules
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
