+++ docs/source/micromind.conversion.rst
@@ -1,0 +1,21 @@
+micromind.conversion package
+============================
+
+Submodules
+----------
+
+micromind.conversion.convert module
+-----------------------------------
+
+.. automodule:: micromind.conversion.convert
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: micromind.conversion
+   :members:
+   :undoc-members:
+   :show-inheritance:
