+++ docs/source/micromind.networks.rst
@@ -1,0 +1,21 @@
+micromind.networks package
+==========================
+
+Submodules
+----------
+
+micromind.networks.phinet module
+--------------------------------
+
+.. automodule:: micromind.networks.phinet
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: micromind.networks
+   :members:
+   :undoc-members:
+   :show-inheritance:
