+++ docs/source/micromind.rst
@@ -1,0 +1,20 @@
+micromind package
+=================
+
+Subpackages
+-----------
+
+.. toctree::
+   :maxdepth: 4
+
+   micromind.conversion
+   micromind.networks
+   micromind.utils
+
+Module contents
+---------------
+
+.. automodule:: micromind
+   :members:
+   :undoc-members:
+   :show-inheritance:
