+++ docs/source/micromind.utils.rst
@@ -1,0 +1,21 @@
+micromind.utils package
+=======================
+
+Submodules
+----------
+
+micromind.utils.configlib module
+--------------------------------
+
+.. automodule:: micromind.utils.configlib
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: micromind.utils
+   :members:
+   :undoc-members:
+   :show-inheritance:
