+++ docs/source/modules.rst
@@ -1,0 +1,7 @@
+micromind
+=========
+
+.. toctree::
+   :maxdepth: 4
+
+   micromind
