+++ tests/test_networks.py
@@ -8,8 +8,8 @@ import torch
 
 
 def test_onnx():
-    from phinet import PhiNet
-    from phinet.conversion import convert_to_onnx
+    from micromind import PhiNet
+    from micromind.conversion import convert_to_onnx
 
     save_path = "temp.onnx"
 
@@ -29,8 +29,8 @@ def test_onnx():
 
 
 def test_openvino():
-    from phinet import PhiNet
-    from phinet.conversion import convert_to_openvino
+    from micromind import PhiNet
+    from micromind.conversion import convert_to_openvino
 
     save_dir = "vino"
 
@@ -45,8 +45,8 @@ def test_openvino():
 
 
 def test_tflite():
-    from phinet import PhiNet
-    from phinet.conversion import convert_to_tflite
+    from micromind import PhiNet
+    from micromind.conversion import convert_to_tflite
 
     save_path = "tflite"
